+++ apps/commerce-api/src/test/java/com/loopers/application/provided/MemberRegisterIntegrationTest.java
@@ -17,7 +17,7 @@ import com.loopers.application.required.MemberRepository;
 import com.loopers.domain.DuplicateMemberIdException;
 import com.loopers.domain.member.Gender;
 import com.loopers.domain.member.Member;
-import com.loopers.domain.MemberFixture;
+import com.loopers.domain.member.MemberFixture;
 import com.loopers.domain.member.MemberRegisterRequest;
 import com.loopers.infrastructure.MemberJpaRepository;
 import com.loopers.utils.DatabaseCleanUp;
